+++ src/parser.c
@@ -667,7 +667,7 @@ static boolean parseMultiAssignmentRest(ParseState *state)
     }
     if (!ParseStateWriteInvocation(
             state, estate.nativeFunction, estate.target, estate.argumentCount,
-            ByteVectorSize(&lvalues) / sizeof(estate) + 1))
+            ByteVectorSize(&lvalues) / (uint)sizeof(estate) + 1))
     {
         return false;
     }
